Extract creator-media storage and serving out of submissions.ts
Moves the creator-media write path (storeCreatorImage/storeCreatorPlaytestShot/
storeCreatorReferenceImages, formatPlaytestContextBlock) and the read path
(GET /api/submissions/:token/shot/:id and /preview/:id) into a new
apps/api/src/delivery/creator-media.ts module. Both sides operate on the same
BuildMediaStore surface and share the mediaByIp/maxMediaPerWindow rate-limit
bucket already used by catalog-routes.ts.

submissions.ts drops from 5015 to 4802 lines; the new file is 259 lines,
0 comment-prose debt.
